requirements.txt
pandas
nltk
fuzzywuzzy
sentence_transformers
scikit-learn
faiss-cpu